New BaseClientImpl Class with implementation of basic HTTP methods added. New BaseClient interface added with basic http methods JsonPlaceholder class devided into Post and Todo classes
