+++ src/main/java/com/example/demo/client/baseClient/BaseClient.java
@@ -1,0 +1,17 @@
+package com.example.demo.client.baseClient;
+
+import org.springframework.http.HttpHeaders;
+import org.springframework.util.MultiValueMap;
+import org.springframework.web.reactive.function.client.ClientResponse;
+
+public interface BaseClient {
+    ClientResponse sendGet(String url, HttpHeaders headers, MultiValueMap<String, String> queryParams);
+
+    ClientResponse sendPost(String url, HttpHeaders headers, MultiValueMap<String, String> queryParams, Object body, Class<?> bodyType);
+
+    ClientResponse sendDelete(String url, HttpHeaders headers, MultiValueMap<String, String> queryParams);
+
+    ClientResponse sendPatch(String url, HttpHeaders headers, MultiValueMap<String, String> queryParams, Object body, Class<?> bodyType);
+
+    ClientResponse sendPut(String url, HttpHeaders headers, MultiValueMap<String, String> queryParams, Object body, Class<?> bodyType);
+}
